Renamed browser libs
